Refactor files in src root directory
