+++ src/constants.js
@@ -3,3 +3,5 @@ export const NOT_AUTHENTICATED = 'NOT_AUTHENTICATED';
 export const GET_MY_RESERVATIONS = 'GET_MY_RESERVATIONS';
 export const FETCH_YACHTS_SUCCESS = 'FETCH_YACHTS_SUCCESS';
 export const FETCH_SINGLE_YACHT_SUCCESS = 'FETCH_SINGLE_YACHT_SUCCESS';
+export const DELETE_SINGLE_YACHTS_SUCCESS = 'DELETE_SINGLE_YACHTS_SUCCESS';
+export const ADD_YACHT_SUCCESS = 'ADD_YACHT_SUCCESS';
